# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Internal trading platform API for Polygon market data and IBKR execution.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List broker accounts
      parameters:
        - name: mode
          in: query
          description: Filter by paper or live environment.
          required: false
          schema:
            $ref: "#/components/schemas/EnvironmentMode"
      responses:
        "200":
          description: Account list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountsResponse"
...
  /positions:
    get:
      operationId: listPositions
      tags: [accounts]
      summary: List open positions
      parameters:
        - name: accountId
          in: query
          required: false
          schema:
            type: string
        - name: mode
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/EnvironmentMode"
      responses:
        "200":
          description: Positions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionsResponse"
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: accountId
          in: query
          required: false
          schema:
            type: string
        - name: mode
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/EnvironmentMode"
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/OrderStatus"
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersResponse"
    post:
      operationId: placeOrder
      tags: [orders]
      summary: Submit an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceOrderRequest"
      responses:
        "201":
          description: Accepted order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/preview:
    post:
      operationId: previewOrder
      tags: [orders]
      summary: Preview a normalized IBKR order payload before submission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceOrderRequest"
      responses:
        "200":
          description: Preview payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderPreview"
  /orders/submit:
    post:
      operationId: submitOrders
      tags: [orders]
      summary: Submit either a normalized order request or raw IBKR order payloads
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/PlaceOrderRequest"
                - $ref: "#/components/schemas/SubmitIbkrOrdersRequest"
      responses:
        "201":
          description: Accepted order payload
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Order"
                  - $ref: "#/components/schemas/SubmitIbkrOrdersResponse"
  /orders/{orderId}/replace:
    post:
      operationId: replaceOrder
      tags: [orders]
      summary: Replace an existing IBKR order using the raw IBKR modify payload
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplaceOrderRequest"
      responses:
        "200":
          description: Updated order snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}/cancel:
    post:
      operationId: cancelOrder
      tags: [orders]
      summary: Request cancellation for an existing IBKR order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelOrderRequest"
      responses:
        "200":
          description: Cancellation request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelOrderResponse"
  /quotes/snapshot:
    get:
      operationId: getQuoteSnapshots
      tags: [market-data]
      summary: Get latest quotes for symbols
      parameters:
        - name: symbols
          in: query
          required: true
          description: Comma-separated ticker symbols.
          schema:
            type: string
      responses:
        "200":
          description: Latest quotes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteSnapshotsResponse"
  /news:
    get:
      operationId: getNews
      tags: [market-data]
      summary: Get latest market news articles
      parameters:
        - name: ticker
          in: query
          required: false
          description: Optional ticker to filter articles by symbol.
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: Maximum number of articles to return.
          schema:
            type: integer
            minimum: 1
            maximum: 50
      responses:
        "200":
          description: Latest market news articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsResponse"
  /universe/tickers:
    get:
      operationId: searchUniverseTickers
      tags: [market-data]
      summary: Search the ticker universe
      parameters:
        - name: search
          in: query
          required: false
          description: Search within the ticker symbol and company name.
          schema:
            type: string
        - name: market
          in: query
          required: false
          description: Restrict results to a market.
          schema:
            $ref: "#/components/schemas/UniverseMarket"
        - name: type
          in: query
          required: false
          description: Optional Polygon ticker type filter.
          schema:
            type: string
        - name: active
          in: query
          required: false
          description: Restrict results to active tickers.
          schema:
            type: boolean
        - name: limit
          in: query
          required: false
          description: Maximum number of search results to return.
          schema:
            type: integer
            minimum: 1
            maximum: 50
      responses:
        "200":
          description: Ticker universe search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UniverseTickersResponse"
  /bars:
    get:
      operationId: getBars
      tags: [market-data]
      summary: Get broker-first historical bars for a symbol or contract
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: timeframe
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/BarTimeframe"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50000
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: assetClass
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AssetClass"
        - name: providerContractId
          in: query
          required: false
          schema:
            type: string
            nullable: true
        - name: outsideRth
          in: query
          required: false
          schema:
            type: boolean
        - name: source
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/BarDataSource"
      responses:
        "200":
          description: Historical bars
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BarsResponse"
  /options/chains:
    get:
      operationId: getOptionChain
      tags: [market-data]
      summary: Get options chain snapshot
      parameters:
        - name: underlying
          in: query
          required: true
          schema:
            type: string
        - name: expirationDate
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: contractType
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/OptionRight"
      responses:
        "200":
          description: Options chain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OptionChainResponse"
  /streams/quotes:
    get:
      operationId: streamQuoteSnapshots
      tags: [market-data]
      summary: Stream quote snapshots over server-sent events
      parameters:
        - name: symbols
          in: query
          required: true
          description: Comma-separated ticker symbols.
          schema:
            type: string
      responses:
        "200":
          description: Quote snapshot SSE stream
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/SseStream"
  /streams/options/chains:
    get:
      operationId: streamOptionChains
      tags: [market-data]
      summary: Stream option chain snapshots over server-sent events
      parameters:
        - name: underlyings
          in: query
          required: true
          description: Comma-separated underlying symbols.
          schema:
            type: string
      responses:
        "200":
          description: Option chain SSE stream
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/SseStream"
  /streams/orders:
    get:
      operationId: streamOrders
      tags: [orders]
      summary: Stream order snapshots over server-sent events
      parameters:
        - name: accountId
          in: query
          required: false
          schema:
            type: string
        - name: mode
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/EnvironmentMode"
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/OrderStatus"
      responses:
        "200":
          description: Orders SSE stream
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/SseStream"
  /streams/accounts:
    get:
      operationId: streamAccounts
      tags: [accounts]
      summary: Stream account and position snapshots over server-sent events
      parameters:
        - name: accountId
          in: query
          required: false
          schema:
            type: string
        - name: mode
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/EnvironmentMode"
      responses:
        "200":
          description: Accounts SSE stream
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/SseStream"
  /streams/stocks/aggregates:
    get:
      operationId: streamStockAggregates
      tags: [market-data]
      summary: Stream broker-derived stock minute aggregates over server-sent events
      parameters:
        - name: symbols
          in: query
          required: true
          description: Comma-separated ticker symbols.
          schema:
            type: string
      responses:
        "200":
          description: Stock aggregate SSE stream
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/SseStream"
  /flow/events:
    get:
      operationId: listFlowEvents
      tags: [flow]
      summary: List options flow events
      parameters:
        - name: underlying
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
      responses:
        "200":
          description: Options flow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowEventsResponse"
...
components:
  schemas:
...
    EnvironmentMode:
      type: string
      enum: [paper, live]
...
    UniverseMarket:
      type: string
      enum: [stocks, indices, fx, crypto, otc]
...
    AssetClass:
      type: string
      enum: [equity, option]
    OptionRight:
      type: string
      enum: [call, put]
...
    OrderStatus:
      type: string
      enum:
        [
          pending_submit,
          submitted,
          accepted,
          partially_filled,
          filled,
          canceled,
          rejected,
          expired,
        ]
    BarTimeframe:
      type: string
      enum: [1s, 5s, 15s, 1m, 5m, 15m, 1h, 1d]
...
    BarDataSource:
      type: string
      enum: [trades, midpoint, bid_ask]
...
    Order:
      type: object
      properties:
        id:
          type: string
        accountId:
          type: string
        mode:
          $ref: "#/components/schemas/EnvironmentMode"
        symbol:
          type: string
        assetClass:
          $ref: "#/components/schemas/AssetClass"
        side:
          $ref: "#/components/schemas/OrderSide"
        type:
          $ref: "#/components/schemas/OrderType"
        timeInForce:
          $ref: "#/components/schemas/TimeInForce"
        status:
          $ref: "#/components/schemas/OrderStatus"
        quantity:
          type: number
        filledQuantity:
          type: number
        limitPrice:
          type: number
          nullable: true
        stopPrice:
          type: number
          nullable: true
        placedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        optionContract:
          oneOf:
            - $ref: "#/components/schemas/OptionContract"
            - type: "null"
      required:
        - id
        - accountId
        - mode
        - symbol
        - assetClass
        - side
        - type
        - timeInForce
        - status
        - quantity
        - filledQuantity
        - limitPrice
        - stopPrice
        - placedAt
        - updatedAt
        - optionContract
    PlaceOrderRequest:
      type: object
      properties:
        accountId:
          type: string
        mode:
          $ref: "#/components/schemas/EnvironmentMode"
        symbol:
          type: string
        assetClass:
          $ref: "#/components/schemas/AssetClass"
        side:
          $ref: "#/components/schemas/OrderSide"
        type:
          $ref: "#/components/schemas/OrderType"
        quantity:
          type: number
        limitPrice:
          type: number
          nullable: true
        stopPrice:
          type: number
          nullable: true
        timeInForce:
          $ref: "#/components/schemas/TimeInForce"
        optionContract:
          oneOf:
            - $ref: "#/components/schemas/OptionContract"
            - type: "null"
      required:
        - accountId
        - mode
        - symbol
        - assetClass
        - side
        - type
        - quantity
        - timeInForce
        - optionContract
    OrderPreview:
      type: object
      properties:
        accountId:
          type: string
        mode:
          $ref: "#/components/schemas/EnvironmentMode"
        symbol:
          type: string
        assetClass:
          $ref: "#/components/schemas/AssetClass"
        resolvedContractId:
          type: number
        orderPayload:
          $ref: "#/components/schemas/JsonObject"
        optionContract:
          oneOf:
            - $ref: "#/components/schemas/OptionContract"
            - type: "null"
      required:
        - accountId
        - mode
        - symbol
        - assetClass
        - resolvedContractId
        - orderPayload
        - optionContract
    SubmitIbkrOrdersRequest:
      type: object
      properties:
        accountId:
          type: [string, "null"]
        ibkrOrders:
          type: array
          items:
            $ref: "#/components/schemas/JsonObject"
      required:
        - ibkrOrders
    SubmitIbkrOrdersResponse:
      type: object
      additionalProperties: true
    ReplaceOrderRequest:
      type: object
      properties:
        accountId:
          type: string
        mode:
          $ref: "#/components/schemas/EnvironmentMode"
        order:
          $ref: "#/components/schemas/JsonObject"
      required:
        - accountId
        - order
    CancelOrderRequest:
      type: object
      properties:
        accountId:
          type: string
        manualIndicator:
          type: [boolean, "null"]
        extOperator:
          type: [string, "null"]
      required:
        - accountId
    CancelOrderResponse:
      type: object
      properties:
        orderId:
          type: string
        accountId:
          type: [string, "null"]
        message:
          type: string
        submittedAt:
          type: string
          format: date-time
      required:
        - orderId
        - accountId
        - message
        - submittedAt
...
    AccountsResponse:
      type: object
      properties:
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/BrokerAccount"
      required:
        - accounts
...
    PositionsResponse:
      type: object
      properties:
        positions:
          type: array
          items:
            $ref: "#/components/schemas/Position"
      required:
        - positions
    OrdersResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
      required:
        - orders
    QuoteSnapshotsResponse:
      type: object
      properties:
        quotes:
          type: array
          items:
            $ref: "#/components/schemas/QuoteSnapshot"
      required:
        - quotes
    NewsResponse:
      type: object
      properties:
        articles:
          type: array
          items:
            $ref: "#/components/schemas/NewsArticle"
      required:
        - articles
    UniverseTickersResponse:
      type: object
      properties:
        count:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/UniverseTicker"
      required:
        - count
        - results
...
    SseStream:
      type: string
...